requirements.txt
pypdf>=4.0.0
PyYAML>=6.0.0
